speed_detection_algo_2/requirements.txt
wheel
setuptools
cmake==3.24.0
icrawler
pymongo
python-dotenv
opencv-contrib-python==3.4.13.47
opencv-python==4.5.3.56
requests
numpy
Flask==2.0.1
Werkzeug==2.0.1
Flask-Debug==0.4.3
matplotlib
pandas
pytest
dlib-bin==19.24.1
gunicorn
Flask-Cors